game_loop修正　＆　関数移動
