+++ srcs/game_loop.c
@@ -13,27 +13,10 @@
 #include "alcu.h"
 #include "game.h"
 #include "get_next_line.h"
-#include "libft.h"
-
-void announce_winner(t_game *game) {
-	printf("ds->map[0]: %d\n", game->map[0]);
-	printf("is_player_turn: %d\n", game->player_turn);
-	if (game->map[0] == 1 && game->player_turn) {
-		ft_putendl_fd("AI's win! you lose.", STDOUT_FILENO);
-	} else if (game->map[0] == 0 && game->player_turn) {
-		ft_putendl_fd("player's win! you won.", STDOUT_FILENO);
-	} else if (game->map[0] == 0 && !game->player_turn) {
-		ft_putendl_fd("AI's win! you lose.", STDOUT_FILENO);
-	} else {
-		ft_putendl_fd("player's win! you won.", STDOUT_FILENO);
-	}
-}
 
 void	start_game(t_game *game)
 {
-	bool is_end = false;
-
-	while (!is_end)
+	while (!is_game_over(game))
 	{
 		print_map(game);
 		if(game->player_turn) {
@@ -44,7 +27,6 @@ void	start_game(t_game *game)
 			// pick_item_ai(game);
 		}
 		game->player_turn = !game->player_turn;
-		is_end = is_game_over(game);
 	}
 	announce_winner(game);
 	free_gnl_buf();
